requirements.txt
streamlit
numpy
scikit-learn
scikit-image
Pillow